+++ Futures/Result.swift
@@ -8,4 +8,22 @@ public enum Result<T> {
   
   /// The result was cancelled
   case Cancelled
+  
+  /// The success value of this result, if any
+  public var value: T? {
+    if case .Success(let result) = self {
+      return result
+    } else {
+      return nil
+    }
+  }
+  
+  /// The error of this result, if any
+  public var error: ErrorType? {
+    if case .Error(let issue) = self {
+      return issue
+    } else {
+      return nil
+    }
+  }
 }